// packages/frontend/src/api/generated/models/ProductsSchema.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * ProductsSchema
 */
export type ProductsSchema = {
    productId: string;
    name: string;
    price: number;
    stock: number;
    description: string;
    image: string;
    category: string;
    categoryId?: string;
    lastUpdated?: string;
};